+++ crates/trading/src/strategy/config.rs
@@ -53,9 +53,10 @@ pub struct StrategyConfig {
     /// The order management system type for the strategy. This will determine
     /// how the `ExecutionEngine` handles position IDs.
     pub oms_type: Option<OmsType>,
-    /// The external order claim instrument IDs.
-    /// External orders for matching instrument IDs will be associated with (claimed by) the strategy.
-    pub external_order_claims: Option<Vec<InstrumentId>>,
+    /// Instrument IDs for which this strategy claims external orders.
+    /// External orders (with an EXTERNAL strategy ID) for matching instrument IDs will be
+    /// associated with (claimed by) this strategy during reconciliation.
+    pub claimed_instrument_ids: Option<Vec<InstrumentId>>,
     /// If OTO, OCO, and OUO **open** contingent orders should be managed automatically by the strategy.
     /// Any emulated orders which are active local will be managed by the `OrderEmulator` instead.
     #[serde(default = "default_false")]
@@ -158,7 +159,7 @@ mod tests {
         assert!(!config.use_uuid_client_order_ids);
         assert!(config.use_hyphens_in_client_order_ids);
         assert!(config.oms_type.is_none());
-        assert!(config.external_order_claims.is_none());
+        assert!(config.claimed_instrument_ids.is_none());
         assert!(!config.manage_contingent_orders);
         assert!(!config.manage_gtd_expiry);
         assert!(!config.manage_stop);